Using the _base_manager for ForeignKey.validate to follow the same logic as the foreign descriptor used.
The manager negotiation which is used to lookup ForeignKey objects
resolves to the _base_manager, however the validation rules for cleaning
a ForeignKey field resovle the manager as the _default_manager.  I
believe this creates inconsistency in what is a Valid relationship.

Consider the example;
I have a large project that looks up model instances using the
_default_manager, if I want to introduce a new field, which by default
should determine the queryset results, e.g. a publish flag, I would
simply update the _default_manager to a manager that excludes records
with this flag set, however this should not restrict ForeignKey
relationships to this instance.

Currently the developer could not simply update the _default_manager,
instead would have to scan their entire codebase for uses of the
_default_manager, or any area where queryset negotiation would normally fall
on a third party library, and explicitly define what manager this part of
the code should use to exclude the relevant instances, because updating
the _default_manager would cause ValidationErrors raised during the
ForeignKey.validate run.

This patch applies the queryset logic that is applied for foreign object
lookup, where it uses the _base_manager instead of the _default_manager.
This way, as the developer, I can decide whether the validation should
include or exclude the instances with the new flag set, and the foreign
object validation logic is consistant with the foreign object lookup.